+++ src/Element/Field/RoughDateRangeField.php
@@ -3,6 +3,7 @@
  * @author    Andrew Coulton <andrew@ingenerator.com>
  * @licence   proprietary
  */
+
 namespace Ingenerator\Form\Element\Field;
 
 use Ingenerator\Form\FormElementFactory;
@@ -38,13 +39,31 @@ class RoughDateRangeField extends AbstractFormField
 
     public function __get($option)
     {
-        if ($option === 'from_field') {
+        if ($option === 'display_value') {
+            return $this->getDisplayValue();
+        } elseif ($option === 'from_field') {
             return $this->from_field;
         } elseif ($option === 'to_field') {
             return $this->to_field;
         }
 
         return parent::__get($option);
+    }
+
+    protected function getDisplayValue()
+    {
+        $values = array_filter(
+            [
+                $this->from_field->display_value,
+                $this->to_field->display_value
+            ]
+        );
+        
+        if ($values) {
+            return implode(' - ', $values);
+        } else {
+            return $this->empty_value;
+        }
     }
 
     public function assignValue(FormDataArray $post)