+++ test/unit/Element/Field/TextareaFieldTest.php
@@ -13,7 +13,10 @@ class TextareaFieldTest extends BaseFieldTest
 {
     public function test_it_is_initialisable_from_schema_array()
     {
-        $this->assertInstanceOf(\Ingenerator\Form\Element\Field\TextareaField::class, $this->newSubject());
+        $this->assertInstanceOf(
+            \Ingenerator\Form\Element\Field\TextareaField::class,
+            $this->newSubject()
+        );
     }
 
     public function test_its_default_html_value_is_empty_string()
@@ -45,16 +48,17 @@ class TextareaFieldTest extends BaseFieldTest
     }
 
     /**
-     * @testWith [{}, ""]
-     *           [{"field": ""}, ""]
-     *           [{"field": "0"}, "0"]
-     *           [{"field": "anything"}, "anything"]
+     * @testWith [{}, "", null]
+     *           [{"field": ""}, "", null]
+     *           [{"field": "0"}, "0", null]
+     *           [{"field": "anything"}, "anything", "anything"]
      */
-    public function test_it_assigns_values(array $values, $expect)
+    public function test_it_assigns_values(array $values, $expect_html, $expect_display)
     {
         $subject = $this->newSubject(['name' => 'field']);
         $subject->assignValue(new FormDataArray($values));
-        $this->assertSame($expect, $subject->html_value);
+        $this->assertSame($expect_html, $subject->html_value, 'Should have html value');
+        $this->assertSame($expect_display, $subject->display_value, 'Should have display value');
     }
 
     /**