+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Unreleased
 
+* Add `display_value` to all fields and use this for rendering in the display-mode
+  templates.
 * Provide helper method on FormElementRenderer to provide the criteria-based highlight classes
 * Include highlight_if and hide_display_if in the field element models
   (we need access to them for display-mode code)